Add CI workflow and modularize manifest code
Introduced a GitHub Actions CI workflow for Rust and refactored the library into modular components. The library's manifest-related code is now housed in its own module, and the package version is updated to 0.2.0.
