+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "opentdf-rs"
-version = "0.1.0"
+name = "opentdf"
+version = "0.2.0"
 edition = "2021"
 
 [dependencies]
@@ -10,3 +10,5 @@ serde = { version = "1.0.214", features = ["derive"] }
 base64 = "0.22.1"
 hmac = "0.12.1"
 sha2 = "0.10.8"
+thiserror = "1.0.67"
+tempfile = "3.13.0"
